[Feat] scaling header UI 수정
